requirements.dev.txt
-r requirements.prod.txt

# Core Django (version pinning)
Django==4.2.5
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
gunicorn==20.1.0
drf-spectacular==0.29.0
dj_database_url>=1.0.0

# Testing
pytest>=9.0.0
pytest-django>=4.11.0
pytest-cov>=4.1.0
factory_boy>=3.3.0
Faker>=38.0.0
httpx>=0.27.0
coverage>=7.2.0

# Type checking
mypy>=1.19.0
django-stubs[compatible-mypy]>=5.1.0
djangorestframework-stubs>=3.15.0
types-redis>=4.6.0
types-requests>=2.32.0
types-django-filter>=23.0.0

# Linting & Formatting (Ruff remplace flake8)
ruff>=0.8.0
black>=24.10.0
isort>=5.13.0

# Debug & Development
ipython>=8.14.0
ipdb>=0.13.9

# Pre-commit
pre-commit>=4.0.0
